backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
httpx
playwright
pyvirtualdisplay
infisical-python